refactor(common): hoist neutral candidate/state machinery out of smda.intel
Move the architecture-neutral function-candidate manager, function candidate,
per-function analysis state and language analyzer into smda.common as shared
bases. smda.intel keeps only the x86 byte-level scans (0xE8 call refs,
prologue seeding, PLT stub chains, x64 .pdata, NOP-aware gap scan) and x86
prologue scoring; smda.aarch64 subclasses the common bases directly and no
longer imports candidate/state plumbing from smda.intel.

The per-arch FunctionCandidate subclass (named via the manager's
CANDIDATE_CLASS factory attribute) is now the explicit function-start scoring
hook: getFunctionStartScore feeds the priority queue (AArch64 stays 0 so
mid-function frame records do not outrank call-ref seeds) and
hasCommonFunctionStart feeds confidence.

Behavior-identical: full SmdaReport output verified byte-identical across all
bundled xored fixtures (aarch64 static golden 278/19881/3525, aarch64 switch
Mach-O, cutwail, bashlite, mirai x64/i386, komplex, pe_export_label_test,
asprox memory dump).

Validation: make lint, make test (534 passed, 1 skipped), fixture report
equivalence diff.
